requirements.txt
numpy
pillow
wandb
matplotlib
scikit-learn
scipy